Project - 21 April - Copy/requirements.txt
Flask==2.2.2
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
pycryptodome==3.17
boto3==1.24.0
python-dotenv==0.21.0
werkzeug==3.1.3